requirements/base.txt
django==5.0.3
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-environ==0.11.2
colorlog==6.8.2
drf-yasg==1.21.7
django-phone-field==1.8.1
Pillow==10.2.0
django-cleanup==8.1.0
